pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gattc"
version = "0.4.0"
description = "BLE GATT schema compiler"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Dmytro Hranchak", email = "kainhurst@gmail.com"}
]
keywords = ["ble", "gatt", "bluetooth", "code-generator"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "jinja2>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-PyYAML",
]

[project.scripts]
gattc = "gattc.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
gattc = ["templates/**/*.j2"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
files = ["src/gattc"]
